restructuring to regular module
