+++ app/models/import_service_config/nemo.rb
@@ -1,7 +1,7 @@
 module ImportServiceConfig
   # methods/configurations for importing data from NeMO Identifiers API
   class Nemo
-    include ImportServiceConfig::Base
+    include ImportServiceConfig
     include Loggable
 
     attr_accessor :project_id
